+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Incremented MSRV to 1.89.
 - Add a constant propagation pass after other mir passes (#439).
+- Add a pass for common subexpression elimination in the constraint graph (#419).
 
 ## 0.4.0 (2025-06-20)